algo_runtime/requirements.base.txt
pyzmq>=26.0
msgpack>=1.0
numpy>=1.26
numba>=0.60
uvloop>=0.21; sys_platform != "win32"
